Make it easier to jump to prefix dirs
When a history entry is the suffix of the given pattern then jump to
that one so that the user doesn't always end up in dirs that are
higher ranked but pattern is only a prefix of them instead of a more
exact match.

Example
=======

History contains two entries: "/home/makkes/a" has a lower rank than
the entry "home/makkes/a/b" because the latter has been visited more
often in the past.

User types: "j /a"

Before this change
------------------

The user would end up in "/home/makkes/a/b" and he would have no
chance to jump to "/home/makkes/a" besides actually typing the full
path (because exact matches are ranked higher than more often visited
non-exact matches).

After this change
------------------

The user would end up in "/home/makkes/a" because the typed pattern
"/a" is a suffix of this history entry and is considered a more exact
match than "/home/makkes/a/b" even though the latter is higher ranked
in terms of visiting frequency.