frequency wins over score when suffixes match
The score applied by the fuzzy library feels a little arbitrary on
certain paths, e.g. when querying "weave-gitops" then
"/home/max/dev/weaveworks/weave-gitops" gets a score of 147620 while
"/home/max/private/e13/infra/weave-gitops" gets a score of 442862.
However, in this case I want the former to be ranked higher if it has
a higher frequency as the latter so I don't have to extend the query
to be more specific.